backend/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
asgiref==3.4.1
    # via django
black==21.12b0
    # via -r requirements.in
click==8.0.3
    # via
    #   black
    #   pip-tools
django==4.0
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   djangorestframework
django-cors-headers==3.10.1
    # via -r requirements.in
djangorestframework==3.13.1
    # via -r requirements.in
isort==5.10.1
    # via -r requirements.in
mypy==0.920
    # via -r requirements.in
mypy-extensions==0.4.3
    # via
    #   black
    #   mypy
pathspec==0.9.0
    # via black
pep517==0.12.0
    # via pip-tools
pip-tools==6.4.0
    # via -r requirements.in
platformdirs==2.4.0
    # via black
pytz==2021.3
    # via djangorestframework
sqlparse==0.4.2
    # via django
tomli==1.2.3
    # via
    #   black
    #   mypy
typing-extensions==4.0.1
    # via
    #   black
    #   mypy
wheel==0.37.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
